+++ src/main/java/com/alarmcontrol/server/ServerApplicationConfig.java
@@ -1,0 +1,35 @@
+package com.alarmcontrol.server;
+
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+import org.zalando.logbook.Logbook;
+import org.zalando.logbook.httpclient.LogbookHttpRequestInterceptor;
+import org.zalando.logbook.httpclient.LogbookHttpResponseInterceptor;
+
+@Configuration
+public class ServerApplicationConfig {
+
+  @Bean()
+  public RestTemplate restTemplate(RestTemplateBuilder builder, Logbook logbook){
+
+    CloseableHttpClient client = HttpClientBuilder.create()
+        .addInterceptorFirst(new LogbookHttpRequestInterceptor(logbook))
+        .addInterceptorLast(new LogbookHttpResponseInterceptor())
+        .build();
+
+    HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
+    requestFactory.setHttpClient(client);
+
+    RestTemplate restTemplate = builder
+        .requestFactory(() -> requestFactory)
+        .build();
+
+    return restTemplate;
+  }
+
+}