+++ src/main/java/com/alarmcontrol/server/notifications/core/messaging/MessageService.java
@@ -1,11 +1,16 @@
 package com.alarmcontrol.server.notifications.core.messaging;
 
 import com.alarmcontrol.server.notifications.core.config.Contact;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
+import org.springframework.util.LinkedCaseInsensitiveMap;
 
 @Service
 public class MessageService {
@@ -19,13 +24,30 @@ public class MessageService {
   public void send(Message message, List<Contact> contacts){
     Assert.notNull(message, "message is null");
 
+    Map<AbstractMessageService, List<Contact>> serviceToContacts =  groupContactsByResponsibleMessageService(contacts);
+
+    for (Entry<AbstractMessageService, List<Contact>> serviceToContactsEntry : serviceToContacts.entrySet()) {
+      AbstractMessageService service = serviceToContactsEntry.getKey();
+      List<Contact> contactsForService = serviceToContactsEntry.getValue();
+
+      sendMessage(message, contactsForService, service);
+    }
+  }
+
+  private Map<AbstractMessageService, List<Contact>> groupContactsByResponsibleMessageService(List<Contact> contacts) {
+    Map<AbstractMessageService, List<Contact>> serviceToContacts = new HashMap<>();
     for (Contact contact : contacts) {
       boolean messageServiceFound = false;
 
       for (AbstractMessageService messageService : messageServices) {
         if(messageService.isResponsible(contact)){
           messageServiceFound = true;
-          sendMessage(message, contact, messageService);
+          if(!serviceToContacts.containsKey(messageService)){
+            serviceToContacts.put(messageService, new ArrayList<>());
+          }
+
+          List<Contact> contactsForService = serviceToContacts.get(messageService);
+          contactsForService.add(contact);
           break;
         }
       }
@@ -36,20 +58,17 @@ public class MessageService {
             contact.getClass().getSimpleName());
       }
     }
+    return serviceToContacts;
   }
 
-  private void sendMessage(Message message, Contact contact, AbstractMessageService messageService) {
-    logger.info("Start sending message '{}' via {} to {} ",
+  private void sendMessage(Message message, List<Contact> contacts, AbstractMessageService messageService) {
+    logger.info("Start sending message '{}' via {}",
         message.getSubject(),
-        message.getClass().getSimpleName(),
-        contact);
+        message.getClass().getSimpleName());
     try{
-      messageService.send(contact, message);
-      logger.info("Following message sent to {}\n: {}",
-          contact,
-          message);
+      messageService.send(contacts, message);
     }catch (Exception e){
-      logger.error("Can not send message due to: {}\n {} \n {}", e.getMessage(), contact, message, e);
+      logger.error("Can not send message due to: {}\n {}", e.getMessage(), message, e);
     }
   }
 }